Added timeouts to all http calls
Last year Avatax went down on Black Friday. This caused a lot of
customers to have long checkout times due to the timeouts coming
from Avalara. Lowering the standard read timeout of the restclient
and adding open timeouts will ensure that connections have time
to be established but if there are problems the customer won't
suffer too much.